requirements.txt
python-telegram-bot==21.1.1
pyyaml==6.0.1
requests==2.31.0
